Created toggle class function
To simplify instances where there is a toggling class I created a toggle class function that takes in the element, the class to add, and the class to remove.

This is different than a jQuery.toggle() because this is swapping a class with another rather than adding it if not there or removing it if present.