Card styles reset immediately to default after delete
When deleting the user-card, you have to reset the state of the card to what you'd like the default display to be, otherwise it passes on its animation to another user card in the dom-repeat after you've deleted this particular card.

Previously there was an attempt to animate the user card and reset it, but that would have required messing with the code in the dom-repeat template. So this particular commit causes the card to snap back to where it needs to be rather than try to animate or bypass dom-repeat recycling.