User-List Code Cleanup
After a lot of changes in the design and functionality of the user-component, a good deal of the CSS and Javascript in user-list was no longer being used. 

So I deleted a lot of it. 

Also depreciated is the use of modes in user-list, which was basically used to expand the new user box. This is now done internally in the "New User" state of the user-component.