bug-fix: user name header shouldn't change till save
When you open an edit user card, it properly displayed Edit: Last, First, but because of data binding to the input form, the header would change along with the input form before actually clicking the save button.

This changes the header to be set with a function that is called whenever this.user is changed.